fix: setup nixDesk as new Btrfs PC
I setup nixDesk. It is encrypted and uses Btrfs.
I tried to automate the install with install-btrfs.sh
This did not work. It failed when downloading to nix-shell which
changed the shell. I also could not get it to work.
I tried to make a keyfile to automate opening the disk after grub.
This did not work ether.